+++ app/models/lets_encrypt_account.rb
@@ -24,7 +24,7 @@ class LetsEncryptAccount < ApplicationRecord
           private_key: private_key,
           directory: LE_DIRECTORY,
           kid: self.account_id.blank? ? nil : self.account_id,
-          bad_nonce_retry: 5
+          bad_nonce_retry: 10
       )
     else
       Acme::Client.new(
@@ -32,14 +32,22 @@ class LetsEncryptAccount < ApplicationRecord
           directory: LE_DIRECTORY,
           kid: self.account_id.blank? ? nil : self.account_id,
           connection_options: { ssl: { verify: false } },
-          bad_nonce_retry: 5
+          bad_nonce_retry: 10
       )
     end
   end
 
   # @see [LetsEncrypt Integration Guide: One Account or Many?](https://letsencrypt.org/docs/integration-guide/#one-account-or-many)
   def self.find_or_create
-    return LetsEncryptAccount.first if LetsEncryptAccount.exists?
+    a = nil
+    can_include = Setting.le_domains_per_account
+    LetsEncryptAccount.all.each do |i|
+      if i.certificates.count < can_include
+        a = i
+        break
+      end
+    end
+    return a unless a.nil?
     a = LetsEncryptAccount.create!
     a.setup!
     a