+++ db/migrate/20220413190431_update_settings.rb
@@ -1,0 +1,11 @@
+class UpdateSettings < ActiveRecord::Migration[6.1]
+  def change
+    Feature.setup!
+    Setting.setup!
+
+    if Setting.belco_enabled?
+      Feature.find_by(name: 'setting_belco').update active: true
+    end
+
+  end
+end